new: Cleanup `WeatherChanger` file.
