+++ src/main/java/dev/salmon/weatherchanger/WeatherChanger.java
@@ -4,6 +4,7 @@ package dev.salmon.weatherchanger;
 import dev.salmon.weatherchanger.command.WeatherChangerCommand;
 import dev.salmon.weatherchanger.config.WeatherConfig;
 import dev.salmon.weatherchanger.listener.WeatherListener;
+import dev.salmon.weatherchanger.util.ForgeHelper;
 import gg.essential.vigilance.Vigilance;
 import net.minecraft.command.ICommand;
 import net.minecraftforge.client.ClientCommandHandler;
@@ -16,8 +17,9 @@ import java.util.Arrays;
 
 @Mod(modid = WeatherChanger.ID, name = WeatherChanger.NAME, version = WeatherChanger.VER)
 public class WeatherChanger {
+
     public static final String NAME = "@NAME@", VER = "@VER@", ID = "@ID@";
-    private static WeatherChanger instance;
+    @Mod.Instance private static WeatherChanger instance;
     private WeatherConfig config;
 
     @Mod.EventHandler
@@ -28,32 +30,16 @@ public class WeatherChanger {
     @Mod.EventHandler
     protected void init(FMLInitializationEvent event) {
         Vigilance.initialize();
-        this.config = new WeatherConfig();
-        this.config.preload();
-        this.registerCommands(new WeatherChangerCommand());
-        registerListeners(new WeatherListener());
+        config = new WeatherConfig();
+        config.preload();
+        ForgeHelper.registerCommands(new WeatherChangerCommand());
+        ForgeHelper.registerListeners(new WeatherListener());
     }
 
-    public WeatherConfig getConfig() { return this.config; }
-
-    /* Used to register forge event listeners */
-    public static void registerListeners(Object... objects) {
-        for (Object o : objects) {
-            MinecraftForge.EVENT_BUS.register(o);
-        }
-    }
-
-    /* Used to unregister forge event listeners */
-    public static void unregisterListeners(Object... objects) {
-        for (Object o : objects) {
-            MinecraftForge.EVENT_BUS.unregister(o);
-        }
-    }
-
-    /* Used to register Forge commands. Should only be used on startup */
-    private void registerCommands(ICommand... command) {
-        Arrays.stream(command).forEachOrdered(ClientCommandHandler.instance::registerCommand);
+    public WeatherConfig getConfig() {
+        return config;
     }
 
     public static WeatherChanger getWeatherChanger() { return instance; }
+
 }
